Bug #32870631: MULTIPLE EQUALITIES CAN GIVE MULTIPLE IDENTICAL HYPEREDGES
If a multiple equality has some parts that induce a hypergraph cycle,
some other parts may cause a duplicate edge. Surprisingly enough, it
seems the rest of the join optimizer handles this gracefully, but it's
still good to have a graph with no duplicates.

This is actually triggered by an existing unit test, we just didn't notice it when writing it.

Change-Id: I5995301c9d6228596b4256723e71eb588ee24726