Bug #32870804: SUPPORT FLATTENING INNER JOINS INTO MULTI-JOINS
Joins in the hypergraph optimizer are generally binary-only,
for simplicity and regularity. However, there are situations where
it would be useful to have generic N-way inner joins, since it makes
pushing more flexible.

In particular, there are some situations where we'd give up pushing
a condition (our existing rewrite code couldn't see deep enough to
make it pushable), and then later make it a cycle-inducing condition
instead. This would cause us to have an unneeded Cartesian product
in the hypergraph; it would in practice never be used in the final
plan, but it would cause us to evaluate somewhat more subplans than
strictly necessary.

We only support this form during pushdown, and we sometimes unflatten
temporarily when doing our associativity rewrites (which are now much
less applicable, as they're only needed for dealing with outer joins).
Thus, it's perhaps not as integrated as we'd like, but it still gives
good gains in some queries (including one existing test case).

Since the join tree moves around a fair bit, so does a number of
test cases. We add a new node sorting function to make them a bit more
robust, and then mostly adjust some indexing in tests to show that they
haven't really changed.

Change-Id: I81a82638769a67172ec51d2db1853ce20f114aab