Bug #32773801: NOT IN CLAUSE GET WRONG RESULT
Prior to running a subquery in Item_int_optimizer, we'd reset the
NULL value flag (set during execution) so that it wouldn't stick.
However, if the subquery was cachable, we should also keep the NULL
flag, so don't reset it.

This is a regression that was made during the split of the index
subquery engines; we just restore the logic.

Change-Id: I909b01ee5db7db1d340c8325e0b508820605ce56