Bug#32788564: TOO MANY COLUMNS ON PARTITIONED TABLE MAKES DATABASE BROKEN
Problem
-------
The server crashed while truncating a partitioned table after it failed
to add number of columns more than the maximum number of columns allowed
in a table.

This happened because the server did not validate the number of columns
being added using ALTER TABLE ... ADD COLUMN to validate the number of
columns in the table at the right place.

Solution
--------
Check and validate that the number of columns being added does not exceed
the maximum number of columns allowed in a table and deny the operation in
such a case.

RB# 26382
Reviewed by: Mayank Prasad <mayank.prasad@oracle.com>