Bug#32381666 MYSQL-8.0 (8.0.24): MYSQLD CRASHED DURING BACKUP WITH ERROR "ASSERTION `M_TOTAL_
Fix to convert 32-bit sizes to 64-bit sizes

The event buffer 'total' size is increased whenever the event buffer
is expanded. This results in an overflow if the new total is greater
than UINT32_MAX. Due to this overflow, a very low value is saved
as the new total, and this logic error is detected upon comparing
the total size to the free space. This issue can be hit if the
event buffer size reaches 2^32 = 4GB.

This is fixed by changing the maximum and total values to be stored
in 64-bit types. Thus the threshold for event buffer size overflow
is increased to 2^64 = 16Exabytes.

A new logevent EventBufferStatus3 is added with 64-bit max, used and
alloc'ed eventbuffer bytes. A new report reason EVENTBUFFER_USAGE_HIGH
is added so that a logevent is generated when eventbuffer usage
crosses 4GB.

An autotest test which used EventBufferStatus2 is changed to parse
and print EventBufferStatus3 logevents. An error insert is added to
simulate 64-bit max, used and alloc'ed bytes, and a new autotest
test uses the error insert to test EventBufferStatus3 with 64-bit
values.

Approved-by: Maitrayi Sabaratnam <maitrayi.sabaratnam@oracle.com>
Change-Id: I96593ac64dfb2933056fe9b9d283480e0806ce22