Bug #32644631: PRELIMINARY FIXES FOR WL #14419 [fix sortahead, noclose]
Fix miscounting of the number of orderings if we had duplicates;
this could cause some sortahead orderings not to be picked up
by the join optimizer.

In practice, this seems hard to trigger in the current code;
the easiest way is by adding many windows with identical ORDER BY
specifications when that time comes.

Change-Id: I6d63f086fcaa85ba083c38a80ed1373a5d62ae68