Bug#32597017: CSV ENGINE (TINA) DOES NOT HANDLE JSON TYPE
Reading JSON values from tables stored with the CSV engine fails with
an error such as:

ERROR 3144 (22032): Cannot create a JSON value from a string with CHARACTER SET 'binary'.

The reason is that the CSV engine uses Field::store() to copy a value
from the table to the record buffer, and it always passes
my_charset_bin as the character set argument to Field::store().
Field_json::store() has an explicit check for my_charset_bin, and
raises an error if this character set is given.

Fixed by passing the actual character set of the column instead of the
character set of the buffer holding the data (which is always binary).

Change-Id: I446361d1c996d5babdc68f8b133faea3e2f135b7