Bug#32863713: RECENT REGRESSION: CTE CRASH IN               QUERY_BLOCK::MASTER_QUERY_EXPRESSION
To clone a derived table expression, we re-parse it. To re-parse the
expression, a new query block is created and used temporarily.
However, as part of an earlier fix, the query block to be used for parsing
was set to the query block provided in the context (derived table query
block or in case of a view ref, the merged derived table query block).
But, the query expression to use was still the query expression created
temporarily for parsing. Because of the mismatch between the query block
and the query expression that it belongs to, parser tries to access
an invalid query expression resulting in server exit.

We now set the query expression along with the query block to parse
the derived table expression.

Change-Id: If65b5b2d359d1c35c7e3984f301cb089e54e65d1