BUG#32892977: STOP REPLICA IO_THREAD ON CHANNEL WITH ACF ENABLED ALSO STOPS SQL THREAD
When SOURCE_CONNECTION_AUTO_FAILOVER is enabled on a channel, the IO
thread of that channel is monitored for connection failures,
retrying its connection the configured sources if needed.
`STOP REPLICA IO_THREAD` command was not taking into consideration
the monitor thread, that is, if the monitor thread should also be
stopped if it is the last channel being monitored.

To solve the above issue, the `STOP REPLICA IO_THREAD` command
implementation now does stop the monitor thread if needed.
It was corrected the constant that identifies the monitor thread to
avoid clashes with other threads.

RB: 26416