Bug#32424653 PARALLEL 'CREATE USER' HANGS ON BOTH MYSQLDS
When the NDB storage engine is active, all SQL statements that involve
users and grants are evaluated to determine if they effect any users
with the NDB_STORED_USER privilege. After evaluation, some statements
are ignored, some are distributed to all mysql servers as statements,
and some are distributed to all mysql servers as a request to read and
apply a user privilege snapshot that is stored in the ndb_sql_metadata
table. Statement updates are limited to one SQL statement, but snapshot
updates can involve a long list of changes -- up to seven statements
per user for multiple users.

Until now, both statement and snapshot distribution were performed
asynchronously and without any locking. This is noted in the manual.

The problem report here shows that simultaneously creating a user and
then granting NDB_STORED_USER, from two different mysql servers, can
cause the servers to hang. This is because waiting for any lock in the
NDB binlog thread while managing distributed users can easily lead to
deadlock, and the thread was waiting for an exclusive lock on the local
ACL cache.

Therefore this change introduces explicit locking around NDB_STORED_USER
snapshot updates. Snapshot distribution is now performed while holding
a cluster-wide read lock on one row of ndb_sql_metadata. A thread that
fails to obtain this lock in the first try will quickly fail to
distribute the ACL change, raise a warning message for the user, and
prevent the deadlock.

Change-Id: Ib357e2eeeea409e0a551c1aa44bcd807e0a28f9e