Bug#32884066 SOCKETCLOSEONMETADATAAUTHFAILTEST FAILS ON WINDOWS
The test is failing with wait_for_port_not_available(port)

The reason for that is the mock server getting stuck on trying
to log something to the pipe that is full.

This patch changes the configuration of the mock_servers in the
component tests to log to the file instead of the console.

RB: 26404