ug#25449055 NDB_RESTORE TRANSLATE FRM FOR USER DEFINED PARTITIOING TABLES
The ndb_restore utility has code to "translate frm" when restoring a
table from backup which has user defined partitioning describing which
nodegroup each partition should use.

When the backup is restored to a cluster with different nodegrups using
the --ndb-nodegroup_map= argument the "translate frm" code tries to
replace all occurences of NODEGROUP=<old> with NODEGROUP=<new> in the
frm blob attached to the table. Unfortunately this code seems to have
been broken since when it was added in 2006. There are a few problems:
 1) the function translate_frm() returns error when a call to malloc()
    suceeds, it should have checked if the malloc() returned NULL.
 2) crash occurs when search_replace() calls copy_bytes() due to
    segmentation fault.
 3) most of the allocated memory in translate_frm() is not released
    when the function returns.

The purpose of replacing the NODEGROUP= values in the frm blob was to
allow the MySQL Server to open this table to get "proper" values in the
"nodegroup" property of the partition syntax but many implementation
details both in NDB and MySQL Server has since changed and this is not
longer required.

Fix by removing support for mapping nodegroups while restoring backups
to NDB. Leave printout stating that argument value is ignored, this is
to be nice should someone be using the functionality although it never
worked.

Change-Id: I889ddd754755ed59aa2c02004da532761291e9cb