Bug #32644631: PRELIMINARY FIXES FOR WL #14419 [functional dependencies, noclose]
Mark functional dependencies of the type {a} → f(a) as always active,
because they are. This breaks certain expressions involving window
functions, however, so we need to not generate such FDs for them.
(This is true whether always_active is true or not; they simply
do not induce such FDs, since the window functions can depend on
much more state than their immediate arguments.)

Change-Id: I3b66c0dd836a2d3908366a010fba6462b64c55d4