Bug #32907274: COMPILE AS C++17
Now that all of our non-EOL platforms are C++17-capable, turn on the
appropriate flags to compile as C++17, and update compiler checks
for new minimum versions (including switching from Developer Studio
to GCC on Solaris).

After this, the new minimum compiler requirements are:

  - Solaris: GCC 10
  - macOS: XCode 10
  - Linux: GCC 7.1 or Clang 5

Newer versions than the listed ones are, of course, allowed.

Change-Id: I8b26ed85bf37d0b4a88dd35b75b94494662e5f9f