Bug #32908711: HYPERGRAPHSECONDARYENGINETEST.SIMPLEINNERJOIN FAILS ON 32-BIT LINUX
Make it slightly cheaper to return a row from a hash table
than the actual probe lookup; with the current formula, we would
consider (a HJ b) as the same cost as (b HJ a), which would make it
a question of floating-point roundoff which one was preferred when
secondary engine was in use (otherwise, we'd always pick the smaller).
On 32-bit, this would cause a unit test to pick the wrong way due to
80-bit floating-point temporaries.

The number of 0.07 is not tuned at all; it is merely to keep
all the tests passing.

Change-Id: Idefe259fb45de4e80a61358ece008330a752cda0