WL#14628: Step 7: Translate thread stage stored in THD
WL#14628: Compatibility option to enable old replication terminology
in P_S instrumentation

Intercept access to THD::proc_info where it is necessary to convert
the name according to @@session.terminology_use_previous.
This makes INFORMATION_SCHEMA.PROCESSLIST and SHOW REPLICA STATUS show
results according to @@session.terminology_use_previous.

Many places in the code should not translate the string, because they
just save the old stage in order to revert it later.  So we introduce
a new member function that translates the string, and uses it only
where needed.

The existing terminology_use_previous framework assumes that
translated strings have a prefix, corresponding to the
performance_schema instrumentation class, for instance, "stage/sql/".
However, this prefix is added only in performance_schema tables, and
not in INFORMATION_SCHEMA.PROCESSLIST or SHOW REPLICA STATUS.
Therefore, we add an argument to
terminology_use_previous::lookup that allows the caller to
specify that the given string does not have a prefix and the return
value should not have one.

RB:26458