Bug #32889491: INNODB FULLTEXT SEARCH WITH NONAGGREGATED COLUMNS CRASHES MYSQLD
If we have implicit grouping on a single full-text-searched table,
force inserting a temporary table to materialize MATCH() temporaries
before they hit AggregateIterator. This is because AggregateIterator
will want to save and restore the rows it gets in, but cannot properly
include the FTS information, as it is hidden.

Also remove the unused and misleading TABLE::fulltext_searched
member (the real information is in TABLE_LIST).

Change-Id: I805dc6c22dca0f50c0111bd77f72730376266d9d