Bug#31633859: IN GR ENVIRONMENT WITH POOR RESOURCES, EVENT_SCHEDULER DOSEN'T WORK PROPERLY.
When the event scheduler tries to run an event while
SUPER_READ_ONLY is active, it will fail. This is a
side-effect of its trying to update the event's
"last active" timestamp in the mysql schema (which
is rejected in read-only mode).

In response to this inability to process the event
queue, the scheduler will shut down (and note that
fact in the error log).

It is possible to detect this situation in the error log

SELECT error_code,data
  FROM performance_schema.error_log
 WHERE error_code='MY-010045';

or e.g. by checking for the scheduler thread in the
performance schema

SELECT COUNT(thread_id)>0
  FROM performance_schema.threads
 WHERE name='thread/sql/event_scheduler';

This can however be cumbersome in situations where
the server starts in read-only mode (and the event
scheduler therefore shuts down), and then enters
read/write mode once a primary is selected.

As a convenience we therefore now restart the
event scheduler when super_read_only mode is
exited and the event scheduler is configured
to be ON, i.e. once we are able to do so (again),
we produce the state ("event scheduler is running")
that the user requested all along.

Change-Id: I6b8c89eeb2a6c2186f6020cfb95ce344642628a9