Bug#32630875 PAGE TRACKING ABORTS THE SERVER DURING ITS RECOVERY
Issue
=====
During the recovery stage of the server we start the recovery of page
tracking system even before we start the crash recovery process. We
expect this process to be non-blocking. But recently it was seen with a
customer issue that the page tracking system could block the server from
starting if it's not able to read any page tracking file at a particular
offset.

Fix
===
If the file offset we want to read is not present then simple return an
error.

Reviewed-by: debarun.banerjee@oracle.com
RB: 26212