Bug#32905044: CRASH AT CONDITION_PUSHDOWN::REPLACE_COLUMNS_IN_COND DURIN               RQG CONCURRENCY RUNS
An earlier fix (Bug#32324234) made sure that a view reference needs to
be cloned before a condition is pushed down to the derived table.
To resolve the expression that was cloned, merged derived table's
context (query_block->first_context) was used. However, an Item_view_ref
object could also be created when a natural join column is created. For
this case, while resolving, we cannot be using the merged derived table's
context.
Also, query_block->first_context is updated when natural join is present
in the query. As a result, we can no more rely on first_context to give
the correct context for the merged derived table.

Solution:
While creating field translations for the table that is getting merged,
we capture the merged derived table's context. This context is then
saved to each of the Item_view_ref object created. We then use it
when view ref object is cloned.

Change-Id: Ia60c49bf7e8393f2cda5b3ff9be223178e0f7f4a