Bug#32799797: Assertion "handler..ha_release_auto_increment"
When a prepared statement is prepared against a persistent table
and bound against a temporary table for the first execution
(the test case contains a few statements that triggers this scenario),
the preparation leaves a dangling pointer (table in Query_result_insert)
which is accessed in abort_result_set() after a failed execution.

The solution to this is to call the cleanup() function for class
Query_result unconditionally after preparing a prepared statement.
(It should not be done for regular statements and stored routine
statements where the preparation is part of an execution.)

Reviewed by: Jan Wedvik <jan.wedvik@oracle.com>

Change-Id: Iaf85d507585f965d8b37c30c9d5ebdb44268a065