BUG#32798287 PROBLEM WITH UNINSTALL WHEN --REPLICATION-OPTIMIZE-FOR-STATIC-PLUGIN-CONFIG=ON
Description
-----------
When shuting down a server with
`replication_optimize_for_static_plugin_config` enabled, the server is
unable to unload any replication plugin that has been loaded prior to
setting the variable. The server will eventually shutdown but two messages
will be issued to the error log: a warning stating that a plugin is going
to be forcelly unloaded; an error stating the amount of references to the
plugin.

Analysis
--------
When `replication_optimize_for_static_plugin_config` is enabled, the
replication plugin infra-structure holds a plugin reference to every
replication plugin, that is only dereferenced when the option is
disabled. This prevents any given replication plugin to be unloaded and,
thus, allowing to skip some locks related to plugin management, which is
the purpose of enabling the option in the first place.

However, this is not optimal when unloading the plugins at server shutdown.

Fix
---
At server shutdown disable the option and dereference any replication
plugin references stil held.

Reviewed-by: Nuno Carvalho <nuno.carvalho@oracle.com>
Reviewed-by: Pedro Gomes <pedro.gomes@oracle.com>
Reviewed-by: Pedro Ribeiro <pedro.a.ribeiro@oracle.com>
RB: 26340
RB: 26341