BUG#29748305 CAN'T CONNECT MYSQL CLIENTS TO SLAVE IN DEBUG MODE
Issue:
======
In order to connect a client to the slave, we use $MYSQL
--defaults-group-suffix= . But when running mtr --debug,
$MYSQL expands to '/path/to/mysql --debug', so the full
command line becomes '/path/to/mysql --debug --defaults-group-suffix= ',
and this is invalid because '--defaults-group-suffix' must be the first
command line option.

Fix:
====
Instead of using $MYSQL use $EXE_MYSQL for a path and newly
introduced variable $MYSQL_OPTIONS for options.

Change-Id: I6b61a59c6dd20a28f1d2cf541755257efbcb344b