Bug#32813554: Segv in dict0dd.cc:4184 of MySQL server
This problem may occur if we have a common table expression that is
used more than once in a statement, and at least one of the uses is
within a subquery that is removed during resolving because it is
always false (or always true).

At this point in the resolving, handlers have been assigned to all
TABLE objects of the CTE, but none of the TABLEs are yet opened.
When removing the subquery, a CTE reference may also be removed by
calling close_tmp_table() and free_tmp_table() on the TABLE object
representing this CTE reference. However, since tmp_handler_count
is zero, we will actually discard the entire CTE.

When the CTE is actually opened during execution, internal data
structures are inconsistent and the execution fails.

The solution to the problem is to realize that opening a temporary table
is a two-stage process: 1: allocate a handler object and 2: create
the table contents/open the table. By using two counters to track
number of handler counts and open counts per TABLE_SHARE, we are able
to more precisely track the state of the temporary table.
The two counters are named tmp_handler_count and tmp_open_count.
tmp_handler_count is incremented when assigning a handler object in
setup_tmp_table_handler(), and tmp_open_count is incremented in
open_tmp_table(). Both counters are decremented in close_tmp_table(),
if necessary.

Change-Id: I51d9e5a090030e9ee36ca1b051d0abe3b49ac29e