Bug #32888982 EXPORT DATA OF DOUBLE FIELD FAILED TO IMPORT WHEN IT IS DOUBLE'S LOWER BOUNDARY
When converting -DBL_MAX = 1.79769313486231570814527423731704357e+308
to string and then back to double, we get -1.797693134862316e308
which is rejected as out of bounds.

This is a regression from

Bug #28691605 ASSERTION `NUM_CODEPOINTS >= SCANNER.GET_CHAR_INDEX()' FAILED.

Which changed the usage of 'my_gcvt' to prefer 'e' (scientific)
notation when converting a double value to string.

The fix is to avoid rounding errors in the string representation of
double values by increasing the 'width' argument of my_gcvt.

Change-Id: Ia2fddc02b6b0971b4f80b0dcb088d1150161ff72