Bug #32965326 OUT OF THREAD STACK IN CREATEITERATORFROMACCESSPATH (AGAIN)
The patch for
Bug #32934187: HYPERGRAPH: CRASHING IN ITEM_IN_SUBSELECT::EXEC
increased stack usage for CreateIteratorFromAccessPath()
and we have queries which run out of stack space when building with
clang debug and no optimization.

The fix is to increase the value of DEFAULT_THREAD_STACK, and pick a
value which works for all mtr tests and for all known/supported
compilers.  This also simplifies handling of the system variable
"thread_stack", and several -master.opt files of mtr can be removed.

For two test cases, we add a -master.opt file, keeping the old value
for thread_stack, since the test cases are flaky (different behaviour
for different values provided, and actual stack space used).

Previous default value for thread_stack was 280K for most platforms.
New default value is 1M for all platforms.

Change-Id: I101c3959e831bd85ef11dd8ac82ed4c00da1cdb0