Bug#32759421:SEMI SYNC CONNECTION IS OFTEN DROPPED WHEN USING COMPRESSED PROTOCOL
Description:
An error in the communication between source and replica, while using semi-sync
and classic protocol compression, breaks the communication and forces the
link between source and replica to be re-established at every event that needs to be sent.

Analysis:
The replica semi-sync plugin doesn't properly clean up the NET object after sending
reply to the source. The NET packet counters(pkt_nr and compress_pkt_nr) are therefore
left with values != 0. The next time when NET object is used to read a message from the
source it will find a mismatch between the packet serial number received and the
NET object packet counters, assuming it got packets out of order.
Due to this error the connection will be severed.

Fix:
Execute `net_clear` after finishing sending the replica ack message when using the compressed protocol.

Reviewed-by: Pedro Figueiredo <pedro.figueiredo@oracle.com>
Reviewed-by: Sven Sandberg <sven.sandberg@oracle.com>
RB: 26298