Bug #32593352 DDL HANGS IF SCHEMA OP EVENT IS LOST
Problem
=======
If the ndb_schema operation is lost somehow before the coordinator can
handle it, the client, which logged the schema operation is stuck
waiting forever for the coordinator to complete or abort the schema
operation.

Description
===========
Whenever there is a schema change, ndb schema distribution client
running in the user thread creates a schema object describing the
change, pass it to all the participants and the coordinator in binlog
thread and waits for the coordinator to complete the schema
distribution across all the mysqld connected to the  cluster. Upon
completion of the schema distribution or when ndb_schema_dist_timeout
expires, coordinator will wake up client thread by signaling the
conditional variable on which the client thread is waiting on. But, if
suppose the coordinator did not receive the schema object, then the
coordinator will lose the track of the current schema operation and it
can't check for the ndb_schema_dist_timeout or wake up the client
thread because all the required information for performing above tasks
were stored only in the schema object and this for ever leaves the client
thread in the waiting state.

Fix
===
Now, as part of the fix, the client thread is also given access to the
ndb_schema_dist_timeout timer. So, even if the coordinator misses the
event, client thread will detect the timout and will come out from the
waiting state.

Change-Id: I4c2ec0e9a42b7cfc4baf3180c51ea0fc81049807