Bug #33118309 IMPLEMENT INNODB LIBRARY FOR LARGE-PAGE SUPPORT
This patch introduces library functions that are designed to handle
huge-/large-/super-page aligned allocations. Different operating systems
have different names for such pages. Normally, most CPUs support and default
to 4K pages as the smallest memory-addressable unit but they also provide
support for addressing the memory in much larger units, e.g. 2MB, 4MB,
1GiB etc. How much and how many of these units are supported varies from
one CPU to another.

Huge-pages allow to reduce the costs which are imposed by the virtual
memory management routines in the OS kernel and MMU in the CPU. In particular,
huge-pages contribute to reducing the TLB misses and therefore also reducing
the number of soft page-faults. Given that huge-pages are not subject to
the paging algorithms, hard page-faults are impossible to run into. Soft
page-faults are carrying the cost of having to access several memory data
points until the page-walker does not find the right match in the
page-tables. Page-table is generally implemented as a radix-tree. This
practically means that for one virtual-to-physical memory address
translation, CPU will need to traverse through several other ones. If
those ones aren't in the L1/L2 cache, that could very well mean the price
to pay is going to be the latency of at least one round-trip to RAM.

In contrast to the functions implementing huge-page aligned allocations
provided by Unix and/or Windows, these ones have built-in support for
memory instrumentation that is implemented on top of PFS, a performance
schema storage engine. Memory instrumentation is an opt-in feature.

Set of new library functions are:
 * ut::malloc_large_page_withkey
 * ut::malloc_large_page 
 * ut::free_large_page
 * ut::large_page_allocation_size

Several platforms are supported by this library:
 * Linux
 * OSX
 * Solaris 
 * Windows

This set of library functions will allow for more elegant implementation
of the follow-up library whose goal is to allocate memory through
huge-page support but which is able to gracefully fallback to 4K page-aligned
allocations in the event of being unable to back the allocation up from
huge-pages (e.g. huge-page exhaustion).

Reviewed-by: Marcin Babij <marcin.babij@oracle.com>
RB: 26536