WL#14388 Push condition referring fields from an SPJ ancestor table
Remove ancestor restruction imposed by 'firstMatch-return'.

The old mysql server join executor (pre iterator executor) had
a 'firstMatch return' table which the executor jumped out to
when a firstMatch had been found. That mechanism also affected
the access pattern of the pushed join result set, which in turn
required is to enforce that the firstMatch return table was the
parent of the firstMatch'ed table.

This join executor has now been entirely replaced with the
iterator executor, which also means that we do not have to
enforce this anymore.

This patch remove this restriction, or replace it with a less
restrictive limitation that there should be no index-scans inbetween
he ancestor candidates. (Note that there are several places
with similar scan-restriction limitations related to pushing
of outer- and semi-joins.)

Query plans are changed as the result of this. Note that in the
changed query plan the firstMatch'ed semi-join are now allowed
to refer parent tables 'higher up' in the pushed query. This is
an advantageous change as it allows is to earlier detect the
(lack of) a firstMatch for a particular row, and thus eliminate
it from the join processing earlier.

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>