WL#14388 Push condition referring fields from an SPJ ancestor table
Deprecate the QueryTree property bits NI_ATTR_PARAMS and PI_ATTR_PARAMS.

The intended usage of that functionality was to allow the SPJ API
to send a client specified parameter along with its interpreter code,
such that the interpreter instruction could refer values from this
parameter, instead of having the values coded in as part of the
interpreter code.

However, this has never been used, and there are no plans to use
it either. Such values has always been coded into the interpreter
instructions when it is generated.

NOTE, not to be confused with the parameters which this WL is all about:
These parameters are constructed at the (SPJ-) data nodes and send
along with the interpreter code to the LDMs.

Also note that the code path from where this patch remove '*tree.ptr++'
was unused upto now (NI_ATTR_INTERPRET and NI_ATTR_LINKED were unused).
As further patches in this series will start using NI_ATTR_LINKED, we
need to clean up this in advance.

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>